# docs/swagger.yaml
basePath: /v1
definitions:
  data.InputAddUser:
    properties:
      roomID:
        type: integer
      userID:
        type: integer
    type: object
  data.InputAuthUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  data.InputCreateRoom:
    properties:
      title:
        type: string
    type: object
  data.InputRegisterUser:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  data.InputRemoveTask:
    properties:
      roomID:
        type: integer
      taskID:
        type: integer
    type: object
  data.Room:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
  data.Task:
    properties:
      done:
        type: boolean
      id:
        type: integer
      room_id:
        type: integer
      title:
        type: string
    type: object
  data.Token:
    properties:
      expiry:
        type: string
      token:
        type: string
    type: object
  data.User:
    properties:
      activated:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  data.UserTasks:
    properties:
      task:
        items:
          $ref: '#/definitions/data.Task'
        type: array
      user:
        type: string
      userID:
        type: integer
    type: object
  main.Error:
    properties:
      error:
        type: string
    type: object
host: localhost:4000
info:
  contact: {}
  description: API BirgeDo
  title: BirgeDo API
  version: "1.0"
paths:
  /addUser:
    post:
      consumes:
      - application/json
      description: Add User to Room
      parameters:
      - description: Input for adding user
        in: body
        name: roomID
        required: true
        schema:
          $ref: '#/definitions/data.InputAddUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.UserTasks'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      security:
      - ApiKeyAuth: []
      summary: Add User to Room
      tags:
      - Room
  /myrooms/:
    get:
      consumes:
      - application/json
      description: '...'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/data.Room'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      security:
      - ApiKeyAuth: []
      summary: Show User Rooms
      tags:
      - Room
  /mytasks/:
    get:
      consumes:
      - application/json
      description: ..
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/data.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Show User Tasks
      tags:
      - Task
  /removeTask:
    post:
      consumes:
      - application/json
      description: Remove task from Room
      parameters:
      - description: Input for remove user
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/data.InputRemoveTask'
      produces:
      - application/json
      responses:
        "303":
          description: See Other
          schema:
            $ref: '#/definitions/data.InputRemoveTask'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      security:
      - ApiKeyAuth: []
      summary: Remove Task
      tags:
      - Task
  /removeUser:
    post:
      consumes:
      - application/json
      description: Remove user from Room
      parameters:
      - description: Input for remove user
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/data.InputAddUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.InputAddUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      security:
      - ApiKeyAuth: []
      summary: Remove User
      tags:
      - Room
  /room/:
    post:
      consumes:
      - application/json
      description: Create Room ...
      parameters:
      - description: Input for create room
        in: body
        name: InputCreteRoom
        required: true
        schema:
          $ref: '#/definitions/data.InputCreateRoom'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      security:
      - ApiKeyAuth: []
      summary: Create Room
      tags:
      - Room
  /room/{id}:
    get:
      consumes:
      - application/json
      description: get room data by id
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/data.UserTasks'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      security:
      - ApiKeyAuth: []
      summary: Show Room Data
      tags:
      - Room
  /task/:
    post:
      consumes:
      - application/json
      description: Create Task ...
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      security:
      - ApiKeyAuth: []
      summary: Create Task
      tags:
      - Task
  /task/{id}:
    get:
      consumes:
      - application/json
      description: Update Task if true to false otherwise
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      security:
      - ApiKeyAuth: []
      summary: Update Task
      tags:
      - Task
  /tokens/authentication:
    post:
      consumes:
      - application/json
      description: Authentication user
      parameters:
      - description: Input for Auth user
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/data.InputAuthUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/data.Token'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      summary: Authentication User
      tags:
      - User
  /users:
    post:
      consumes:
      - application/json
      description: Registaration user
      parameters:
      - description: Input for remove user
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/data.InputRegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/data.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/main.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Error'
      summary: Register User
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    description: OAuth protects our entity endpoints
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"